* **Key Bindings** - now shows conflicting keyBinds * Renamed ***Full Buffs Please*** to ***Common Buffs*** because it gives you Bless/Haste/Displacement/Heroism
