+++ ToyBox/ReadMe.md
@@ -14,7 +14,9 @@
 **Important**: *Make sure you are on the latest version of the game 1.07f or newer*
 
 ### Ver 1.3.18 (Coming Soon)
+* **Key Bindings** - now shows conflicting keyBinds
 * **Bag of Tricks** - Added a ***Buff Like A Godess*** which makes you practically invulnerable
+* Renamed ***Full Buffs Please*** to ***Common Buffs*** because it gives you Bless/Haste/Displacement/Heroism
 * **Loot Colors** - made scribable scrolls show up as uncommon
 * **Enchantment** - Sandal knows you like to hoard loot so he will help you find items with a handy new Search field in the inventory column
 * **Equipment No Weight** tweek is back. It was stolen by quasits who work for the merchant guild...
